requirements.txt
streamlit==1.38.0
langchain==0.3.1
langchain-community==0.3.1
langchain-openai==0.2.0
pypdf2==3.0.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
tiktoken==0.7.0
openai>=1.40.0
